# openapi.yaml
openapi: 3.0.0
info:
  title: fitlife API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0
servers:
  - url: http://localhost:3003/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /entrenador:
    get:
      tags:
        - Entrenador
      summary: Listar todos los Entrenadors
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Entrenadors
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Entrenador
      summary: Crear Entrenador
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Entrenador creado
  /entrenador/{id}:
    get:
      tags:
        - Entrenador
      summary: Obtener Entrenador por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entrenador encontrado
        '404':
          description: Entrenador no encontrado
    patch:
      tags:
        - Entrenador
      summary: Actualizar Entrenador
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Entrenador actualizado
    delete:
      tags:
        - Entrenador
      summary: Eliminar Entrenador
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entrenador eliminado
  /entrenador/{id}/upload-image:
    post:
      tags:
        - Entrenador
      summary: Subir imagen a Entrenador
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
  /plan-entrenamiento:
    get:
      tags:
        - PlanEntrenamiento
      summary: Listar todos los PlanEntrenamientos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de PlanEntrenamientos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - PlanEntrenamiento
      summary: Crear PlanEntrenamiento
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: PlanEntrenamiento creado
  /plan-entrenamiento/{id}:
    get:
      tags:
        - PlanEntrenamiento
      summary: Obtener PlanEntrenamiento por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PlanEntrenamiento encontrado
        '404':
          description: PlanEntrenamiento no encontrado
    patch:
      tags:
        - PlanEntrenamiento
      summary: Actualizar PlanEntrenamiento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: PlanEntrenamiento actualizado
    delete:
      tags:
        - PlanEntrenamiento
      summary: Eliminar PlanEntrenamiento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PlanEntrenamiento eliminado
  /plan-entrenamiento/{id}/upload-image:
    post:
      tags:
        - PlanEntrenamiento
      summary: Subir imagen a PlanEntrenamiento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
...